added better timeout handling to picoscopev2
picoscope uses an autotrigger if the timeout is exceeded, meaning a trace is recorded and cvt is updated even in the event of a timeout. A timer was added that checks the amount of time it takes to get new data to the buffer on the picoscope. If this is larger than the timeout value (minus 200 ms to acccount for variations) the buffer is read, but discarded and no cvt or subscriber updates happen